CC-2190 CC-2215 Restore filter icons
See also commit 09b842eab12adf6b486e4e96443db19d50b72e58
of vocabs-registry, which updated
conf/ANDS-ELDAConfig-template.ttl.sample
and which is necessary for this commit to work: both
that commit and this one must be deployed together.

Restore the "<" and ">" filter icons, because
we can now make them work with date/time values.

This also now works on the /resource endpoint,
because the link now correctly points to
the /concept endpoint.