CC-2561 In HTML results, honour newline characters in data
In HTML pages generated by Elda/SISSVoc, newline characters in the
vocabulary data are sent to the browser as-is, which means that they
are handled by the browser, which renders them as a space.

It seems that several users want to have newline characters shown as
such in the browser.

It's reasonable to honour newline characters that are in the vocabulary
data.

Therefore: add a patch to ands-ashtml-sissvoc.xsl to replace embedded
newline characters with <br /> elements.

NB: we use an XSLT 2.0 feature, so ands-ashtml-sissvoc.xsl now begins
<xsl:stylesheet version="2.0" ...

CC-2561 CC-2583