CC-2369 Show help icons in right-hand panels
There should be help icons (question marks) in the right-hand
panels, which you can click to display/hide some help text. But
they're not shown, because we're using a version of jQuery that
removed the toggle() function used to make them visible.

Since recent jQuery toggle() doesn't do what it used to, replace
its uses with a new function.  Function definition sourced from
https://forum.jquery.com/portal/en/community/topic/beginner-function-toggle-deprecated-what-to-use-instead
and https://jsfiddle.net/s376u4zn/1/ .

Following on: now that the help buttons are visible, if you click
the one for "Sort by", the help text is long, and can extend over
the following "View" box. Add some CSS overrides so that the
question mark icon of the "View" box doesn't stick through the
popup.

RVA-292