CC-2368 Don't break sort panel after removing a sort field
When viewing a Linked Data list page, you can add multiple sort
criteria. If you remove one that's not the first, the display of
sort criteria gets messed up: you get one or more "empty"
criteria added.

The cause is multiple instances of carelessness with the
insertion of commas before and after sort fields.

Add a patch template that is careful to insert a comma when and
only when it is necessary.

RVA-293