CC-2370 Fix less/greater filters better
In CC-2190 I restored the less/greater filter buttons.

It seems more work is needed for these buttons on the item endpoint.

For example:

* Using a vocabulary whose concepts have dcterms created/modified values,
  go to a concept page, e.g., ending in concept.html.
* Click the first concept on the list, to bring up the item endpoint.
* In the "View" box on the left, click "description".
* Now try one of the less/greater filters against either "dcterms created"
  or "dcterms modified". You get an error page.

The reason is that "Patch 7" of ands-ashtml-sissvoc.xsl doesn't manage to
remove the uri=... parameter once you have added extra "bits" to the URL.
In this case, selecting the "description" viewer adds an extra component
to the URL, and the regular expression that looks for the uri=...
parameter no longer matches.

Re-implement the replacement process in two steps:
* Replace the resourceEndPoint with conceptSearchEndPoint using
  fn:replace.
* Remove any "uri=..." parameter using the substituteParam
  template.

RVA-294