CC-2466 devl, vocabs: replace SVC_PREFIX with correct values
Oh dear, when I incorporated the changes made to the template for
RVA-298, I mistakenly added the definitions of the new endpoints
straight out of the template, without replacing SVC_PREFIX with the
correct value for each vocabulary.

Fix all of the specification files so that they use the
vocabualary-specific value.

RVA-298